Fix: 랭킹 기능 최초값 설정 및 랜더링 미 적용 수정
